Create hooks file for gql queries
