+++ src/components/FavoriteDistrictsPage/FavoriteDistrictsPage.js
@@ -3,43 +3,22 @@ import './FavoriteDistrictsPage.scss';
 // import PropTypes from 'prop-types'
 import FavoriteReportCard from '../FavoritedReportCard/FavoriteReportCard';
 import {  useNavigate } from 'react-router-dom';
-import { useMutation, gql } from '@apollo/client';
+import { useAddFavorite } from '../../hooks/useAddFavorite';
 
 
-
-const FavoriteDistrictsPage = ({currentDistrictID, currentUserID}) => {
-    //will this need to hold state [] for the favorites or can we hold this in App?? 
+const FavoriteDistrictsPage = ({currentDistrictData, userData}) => {
     const navigate = useNavigate()
 
-
-    const FAVORITE_DISTRICT = gql`
-  mutation createUserDistrict($userId: Number!, $districtId: Number! ){
-    createUserDistrict(input: {
-        userId: $userId,
-        districtID: $districtId
-    }) {
-        userdistrict {
-            id
-        }
-    }
-}
-`;
-
-const { error, loading, data } = useMutation(FAVORITE_DISTRICT, {
-    variables: {
-        userId: currentUserID,
-        districtId: currentDistrictID
-    }
-  })
-
-
-
+    let districtId = currentDistrictData.data.attributes[0].lea_id
+    let userId = userData.user.id
+    const { error, loading, data } = useAddFavorite(userId, districtId )
 
     return (
         <div className='favorites-container'>
-            FavoriteDistrictsPage WILL EXIST HERE
-            <button className='search-page' onClick={ () => navigate('/home') }>Back to Search Page</button>
-
+            <FavoriteReportCard
+                // data={data}
+            />            
+           <button className='search-page' onClick={ () => navigate('/home') }>Back to Search Page</button>
         </div>
     )
 }
